Commonize the fragmentum
