+++ app/src/main/java/mozilla/lockbox/presenter/FxALoginPresenter.kt
@@ -7,26 +7,19 @@
 package mozilla.lockbox.presenter
 
 import io.reactivex.Observable
-import io.reactivex.disposables.CompositeDisposable
 import io.reactivex.rxkotlin.addTo
 import mozilla.lockbox.action.RouteAction
 import mozilla.lockbox.flux.Dispatcher
+import mozilla.lockbox.flux.Presenter
 
 interface FxALoginViewProtocol {
     val logMeInClicks: Observable<Unit>
 }
 
-class FxALoginPresenter(private val protocol: FxALoginViewProtocol, private val dispatcher: Dispatcher = Dispatcher.shared) {
-
-    private val compositeDisposable = CompositeDisposable()
-
-    fun onViewReady() {
+class FxALoginPresenter(private val protocol: FxALoginViewProtocol, private val dispatcher: Dispatcher = Dispatcher.shared) : Presenter() {
+    override fun onViewReady() {
         this.protocol.logMeInClicks.subscribe {
             dispatcher.dispatch(RouteAction.ITEMLIST)
         }.addTo(compositeDisposable)
-    }
-
-    fun onDestroy() {
-        compositeDisposable.clear()
     }
 }
