+++ app/src/main/java/mozilla/lockbox/view/FxALoginFragment.kt
@@ -1,7 +1,6 @@
 package mozilla.lockbox.view
 
 import android.os.Bundle
-import android.support.v4.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -12,10 +11,7 @@ import mozilla.lockbox.R
 import mozilla.lockbox.presenter.FxALoginPresenter
 import mozilla.lockbox.presenter.FxALoginViewProtocol
 
-class FxALoginFragment : Fragment(), FxALoginViewProtocol {
-
-    lateinit var presenter: FxALoginPresenter
-
+class FxALoginFragment : CommonFragment(), FxALoginViewProtocol {
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
@@ -23,16 +19,6 @@ class FxALoginFragment : Fragment(), FxALoginViewProtocol {
     ): View? {
         presenter = FxALoginPresenter(this)
         return inflater.inflate(R.layout.fragment_fxa_login, container, false)
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-        presenter.onViewReady()
-    }
-
-    override fun onDestroy() {
-        super.onDestroy()
-        presenter.onDestroy()
     }
 
     override val logMeInClicks: Observable<Unit>
