+++ app/src/main/java/mozilla/lockbox/view/SettingFragment.kt
@@ -5,7 +5,6 @@
 package mozilla.lockbox.view
 
 import android.os.Bundle
-import android.support.v4.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -13,21 +12,15 @@ import mozilla.lockbox.R
 import mozilla.lockbox.presenter.SettingPresenter
 import mozilla.lockbox.presenter.SettingViewProtocol
 
-class SettingFragment : Fragment(), SettingViewProtocol {
-    private var presenter = SettingPresenter(this)
-
+class SettingFragment : CommonFragment(), SettingViewProtocol {
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
+        presenter = SettingPresenter(this)
+
         // Inflate the layout for this fragment
         return inflater.inflate(R.layout.fragment_setting, container, false)
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-
-        presenter.onViewReady()
     }
 }
